ryan's 16/05

+++ app/src/main/java/com/fit/quizcrafter/domain/weather/Current.java
@@ -13,20 +13,20 @@ import java.util.List;
 public class Current {
 
     private String observation_time;
-    private int temperature;
-    private int weather_code;
+    private double temperature;
+    private double weather_code;
     private List<String> weather_icons;
     private List<String> weather_descriptions;
-    private int wind_speed;
-    private int wind_degree;
+    private double wind_speed;
+    private double wind_degree;
     private String wind_dir;
-    private int pressure;
-    private int precip;
-    private int humidity;
-    private int cloudcover;
-    private int feelslike;
-    private int uv_index;
-    private int visibility;
+    private double pressure;
+    private double precip;
+    private double humidity;
+    private double cloudcover;
+    private double feelslike;
+    private double uv_index;
+    private double visibility;
     private String is_day;
     public void setObservation_time(String observation_time) {
          this.observation_time = observation_time;
@@ -35,17 +35,17 @@ public class Current {
          return observation_time;
      }
 
-    public void setTemperature(int temperature) {
+    public void setTemperature(double temperature) {
          this.temperature = temperature;
      }
-     public int getTemperature() {
+     public double getTemperature() {
          return temperature;
      }
 
     public void setWeather_code(int weather_code) {
          this.weather_code = weather_code;
      }
-     public int getWeather_code() {
+     public double getWeather_code() {
          return weather_code;
      }
 
@@ -63,17 +63,17 @@ public class Current {
          return weather_descriptions;
      }
 
-    public void setWind_speed(int wind_speed) {
+    public void setWind_speed(double wind_speed) {
          this.wind_speed = wind_speed;
      }
-     public int getWind_speed() {
+     public double getWind_speed() {
          return wind_speed;
      }
 
     public void setWind_degree(int wind_degree) {
          this.wind_degree = wind_degree;
      }
-     public int getWind_degree() {
+     public double getWind_degree() {
          return wind_degree;
      }
 
@@ -87,49 +87,49 @@ public class Current {
     public void setPressure(int pressure) {
          this.pressure = pressure;
      }
-     public int getPressure() {
+     public double getPressure() {
          return pressure;
      }
 
     public void setPrecip(int precip) {
          this.precip = precip;
      }
-     public int getPrecip() {
+     public double getPrecip() {
          return precip;
      }
 
     public void setHumidity(int humidity) {
          this.humidity = humidity;
      }
-     public int getHumidity() {
+     public double getHumidity() {
          return humidity;
      }
 
     public void setCloudcover(int cloudcover) {
          this.cloudcover = cloudcover;
      }
-     public int getCloudcover() {
+     public double getCloudcover() {
          return cloudcover;
      }
 
     public void setFeelslike(int feelslike) {
          this.feelslike = feelslike;
      }
-     public int getFeelslike() {
+     public double getFeelslike() {
          return feelslike;
      }
 
     public void setUv_index(int uv_index) {
          this.uv_index = uv_index;
      }
-     public int getUv_index() {
+     public double getUv_index() {
          return uv_index;
      }
 
     public void setVisibility(int visibility) {
          this.visibility = visibility;
      }
-     public int getVisibility() {
+     public double getVisibility() {
          return visibility;
      }
